Version control requirements_for_test_common.txt
We excluded this file from version control because it’s copied
automatically from utils, so we only need to modify it and track changes
there. In theory at least.

In practice Dependabot cannot:
- cope with having a requirements file which isn’t in the repo
- be configured to only look at certain requirements files

So I think the best thing to do is keep this file in version control. It
already has an automatically generated comment to suggest it shouldn’t
be edited manually.

Dependabot may try to raise PRs against it, but we can resolve these by
updating the upstream version in utils. We can then roll out several
version bumps in a single PR per app, rather than several.