Clean up old and processed unsubscribe requests
The way this works is, every night:
1. A task runs and looks for any services which might have unsubscribe
   requests to clean up
2. For each of these services it kicks off 2 sub tasks to:
    1. Remove requests associated with a report which was processed more
       than 7 days ago
    2. Remove requests received more than 90 days ago

Each of these 2 tasks:
1. Finds any relevant requests
2. Copies them into the history table
3. Deletes exactly the same ones from the main table

We don’t touch the reports. Once all the requests in a report has been
deleted, the report won’t be returned to the admin app (per
https://github.com/alphagov/notifications-api/pull/4170/files).

The downside of this way of doing things is we will create more tasks
than are strictly needed. Most of the created tasks will end up
modifying no data (which is quick at least). However the upside is we
don’t have:
- multiple separate queries (potentially 2 to create the tasks,
  2 to execute them, 1 to clean up the reports) which need keeping in sync
  to avoid requests falling between the cracks.
- large volumes of data being passed around as task arguments