api clients: redesign initialization, replace global instances with LazyLocalGetter instances
using globals is not thread safe and will start to cause
problems in our api clients as soon as they start doing
things that involve persistent state. instead, we need to
use thread-local instances of these api clients, allowing
the internal design of those clients to be unconcerned
with thread contexts.

though we only use threads (or rather eventlet) on a few
of our ecs services, we want to have the flexibility to
make future choices about threading without nasty surprises.

the flask-style `init_app()` pattern is particularly
awkward for keeping thread-local copies, as we need to be
able to construct new thread-local instances on-demand
(because we don't necessarily know how many threads we have
initially and threads can be re-created at any time). the
`init_app()` pattern however rather assumes all
initialization will happen at app-init time. it's also
awkward because it means a client instance has an "existing
but not yet configured/valid" state that makes it hard to
reason about or add sensible type-hints for.

so this moves all initialization to the constructor. that
however means we can't construct an instance until we have
an app instance to construct it with - until then we have
to provide the LazyLocalGetter with a "dummy" factory
function that will raise an exception if access is attempted
too early.

in create_app, these factory functions are replaced with
real implementations that are bound to the app instance
(or its config variables).

an alternative option might have been to keep a
queue.LifoQueue of each client and use a flask hook to
provide each new app context with a lazy-getter that would
assign an instance from that queue (or construct one) and
then have a hook that returned instances to the queue on
app context teardown. however that would probably involve
quite a bit of closure trickery and tie us ever closer to
flask.