test_post_notifications: tighten up test assertions, fix incorrect expected result of document download call
this adds assertions that
document_download_client.upload_document(...) is *not* called
when it's not expected to be. however it finds a case when it
actually *is* called despite the test having `expect_upload`
as False. the only explanation we can come up with is that this
is a mistake, otherwise e.g. `expected_filename` wouldn't have
been specified because it has no use when `expect_upload` is
False.

original author sam suggests it might be a typo too from a
cursory glance.