Only return unsubscribe reports with requests
When we start to delete old unsubscribe requests we will leave behind
the reports to which they were joined.

We don’t want to delete these reports complete, since it’s useful for
us to keep a record of what happened when, and how many requests there
were. We do a similar thing for `Job`s where they stay in the database
for all time.

However it’s not useful to show these old reports on the dashboard,
because if a user clicks through to them they won’t be able to access
the underlying requests.

So this commit modifies the function which returns the list of reports
to the admin app, excluding those which no longer have any requests
attached to them.