Updated phonenumber validation code to use allow_landline flag
In order to migrate phonenumber validation code to use the PhoneNumber utils class
we need to support passing the SMS_TO_UK_LANDLINES permissions as an attribute to
instances of the class so that it can correctly validate depending on the permissions
service has. This is also required to fully migrate all phonenumebe validation logic
away to PhoneNumber objects.