Rewrite unsubscribe requests tests to use a helper
For lots of objects we create reguarly in tests we have a little helper
function that:
- makes the code less verbose
- handles creating the associated objects automatically

We didn’t have this for unsubscribe requests and reports. Now we do.

I’ve also modified some of the tests to assert against actual hard
values, rather than results of queries. This makes them less brittle and
harder to hide bugs in.