Add `created_at` column to unsubscribe request reports
So we can perform actions a certain nuber of days after a report is
first downloaded, we need to know when it was first downloaded. This is
also the same time it gets created. Therefore we can use the pattern of
a `created_at` column.

Need to deploy this as a migration first before adding the column to the
model.