Backfill `created_at` dates for unsubscribe request reports
So that the codebase doesn’t have to deal with null values in this
column we need to make sure every row has a value. The value chosen is
the date when the table was first created. So it’s impossible for any
reports to have this as their real `created_at` date.

Secondly we can modify the table to make the column non-nullable.

We will need one final pull request to update the model to also be
non-nullable.