Refactor how we get fields from letter callback requests
We were returning `page_count` and `status` as a tuple when extracting
the properties from the letter callback request data. Now that we are
going to start using more fields from the data, it makes things a little
clearer and avoids mistakes if we return the data as a `dataclass` and
access the fields we want as properties.