Make process_letter_callback_data task save to letter despatch table
We already have a DAO function for saving to the
`notification_letter_despatch` table but that relied on looking up the
notification by reference. We now use the ID of notifications, so this
creates a similar version of that function that updates by ID instead.

When we get the callback data from DVLA we need to get the date in the
right format and work out whether the letter should count as `sorted` or
`unsorted` from the mailing product and postage. We parse the data to
get the values in the right format, then the task uses these to update
`notification_letter_despatch`.