Allow response with no results for get_sll_notitications_for_service()
This is because we are moving to another way of assessing if there
exists next page of notifications.

So far we have always pulled current and next page, to make sure
next page exists.

This new way will only pull current page, and if that page is full,
it will assume that most likely next page does exist.

This is more efficient, but can occasionally result in us serving
an empty page. For that to work, we need to not serve an error
when we don't find results for notifications.