Start split of get_all_notifications_for_service()
Into two endpoints:
- one that returns notifications batches for SCV reports
- one that returns notifications batches to show on the page

Why is this - having one endpoint with so many options and flags
is confusing. So to simplify - we split it into two leaner
endpoints.

In this commit we split out the new endpoint, but the old one
stays as is. Once admin app knows to use the new endpoint for
CSVs, we will make the old endpoint leaner, too.