remove database tables
now that there's no data that depends on these, we can delete the
tables completely. this was important to do as separate steps, or we
could end up in a case where have have fragments of broadcast services
sitting in our database but without the broadcast-specific tables, which
would leave us unsure as to what those services were.