remove broadcast from template_type
we can't _just_ do the same process as 0467 because there are check
constraints that depend on these. When you say "alter column
template_type set type = new_type using old_type::text::new_type", it
takes that "using" statement as a way of translating the prev enum into
the new enum, in our case just casting to a str as an intermediate step.

however, the template_type column has several check constraints against
it (to say only letters can have a letter_attachment for example).
Postgres isn't smart enough to take the `using` clause and apply it to
these constraints. There are two ways around this:

* You could create a comparator function that tells postgres "if you
  have to compare old_type with new_type, here's a custom function"[^1]
* You can drop the old constraints and recreate them after the migration
  is done

We've gone with the second option as it's a bit simpler and I could copy
and paste from our previous scripts

[^1]: https://stackoverflow.com/a/57952345 was very helpful for this