manually create and delete enum in existing alembic migration
there's a bug/design flaw in alembic where if you autogenerate a
migration script, it'll create the enum implicitly, but won't include it
in the downgrade. This is due to complex reasons that I kinda glazed
over.[^1][^2]

With this migration, if you run downgrade and then upgrade again it
fails due to the enum not being dropped by the downgrade, and thus
already existing for the upgrade.

Lets drop the enum manually in the downgrade script.
create the enum manually (not because we need to, but just to give a bit
of clarity as to what's going on behind the scenes)

[^1]: https://github.com/sqlalchemy/alembic/issues/89
[^2]: https://github.com/sqlalchemy/alembic/issues/868#issuecomment-867914756