ensure we delete the same files from s3 as we do from the database
we were not ordering our queries previously - this meant that when we
fetched 5000 letters to delete from s3, and then later in the task
fetched 5000 letters to move from notifications to notification_history,
we wouldn't be guaranteed the same 5000. With all the thrashing of
the db as we're deleting across every service all at once, it's hard to
reason about what postgres may or may not cache.

if these two queries returned different things, we'd end up in a state
where we may delete things from the database before we've deleted them
from s3 - but without them in the notifications table we've got no way
of being able to find them in s3! There's a 90 day aws lifecycle policy
preventing anything from sticking around for _too_ long but still it's
not great practice.

order both by created_at (asc) to ensure we always delete the oldest
first. this also means if the task fails the oldest things are gone,
which seems like a sensible way to do it.