validate authentication before we check jsonschema validity
previously, if they also have bad json,

if someone has bad auth, that should be the first thing we report on.
this means we can't use the json schema validity decorator (which we
only used in this file anyway). We'd be able to use it if we _also_
had an authentication decorator or middleware, but i figured for just
this one endpoint it's as easy to just move the code back into the
functions. updated the tests since the ones that test if the schema is
valid now also need to have valid auth.