Update DVLA letter callback spec
DVLA have informed us that their `Print Date` variable is actually
`productionRunDate`, and is in local time rather than UTC. This
updates the callback to use the new name and timezone.

We can no longer use the built-in JSON schema validator of
`date-time`, so have to use a custom validator.