add new mock_celery_task fixture that checks params
previously, we'd just use a plain old mocker.patch to mock the
apply_async function for any celery tasks we call. This is great for
letting us assert if the task was called with the args/kwargs we expect,
but falls into the classic Mock trap of not actually guaranteeing that
your task call is correct - it is only comparing against the test, and
not the actual task definition. It relies on developers to manually
update tests if the task definition changes (for example, adding a new
arg/kwarg).

This new fixture can be called with a celery task as the only argument,
and it returns a mocked version of that task's `apply_async` method.

You can then assert if that mock was called with the correct args/queue
as you previously have done - but it will _also_ check that:

* the queue name exists in the config
* any args/kwargs map correctly to the celery task definition
* all args/kwarg values are json serializable