Notification: add functional-dependency multivariate statistics
the intention of this is to allow postgres to gather statistics
that should allow it to make better row estimates and therefore
better query plans.

because sqlalchemy has no support for declarative extended
statistics objects (in the way it has support for e.g. indexes),
rather than just applying migrations and having no record of
the up-to-date expected set of extended statistics for a table,
this adds an __extended_statistics__ attribute to the model
class, which is compared in a new test against the extended
statistics objects found in the database for that model's table.
it will fail if they disagree, hopefully preventing anyone from
adding a migration changing extended statistics objects without
reflecting it in the model or making changes to the model without
realizing these changes need to be explicitly applied in a
migration.

out of scope: autogenerating migrations or inspecting any
expression-statistics.