Add functionality to try_parse_and_format_phone_number method to exclude country code when a flag is passed
In certain situations we don't want to the numbers country code to be pre-pended to the formatted number. A good example is when we are wanting to query the db with the LIKE or ILIKE operators. Formatting with the country code can produce unexpected formatting when searching on a number fragment, so searching is easier if this is just removed.