fetch_notification_status_for_service_for_today_and_7_previous_days: perform Template join late
we don't need to join to Template until after the aggregation is
done, which reduces the size of the join and means we don't have
to include all the desired columns from Template in group_by. this
should be a bit faster and is less likely to confuse the planner.