Notification: add ix_notifications_failed_service_id_composite
notifications with failed status are comparatively rare and tend
to sit in cold regions of the table. added to them being composed
of several different individual status types (which rules out a
lot of standard index operations), postgres finds it hard to query
these efficiently.

this is a partial index only covering
NOTIFICATION_STATUS_TYPES_FAILED (which conveniently at time of
writing is the same set of status types requested together by
admin), so it should be a small index that should allow such a
query to completely disregard the notification_status clause and
start reading out the results in created_at order with very little
more filtering.