get_notifications_for_service: add with_personalisation argument
this allows a caller to defer the _personalisation column, which
is generally the biggest field in a row. the field is rarely used
in bulk requests and including it inflates the size of intermediate
results in the query plan, increasing the size of e.g. external
sorts. omitting it will probably also reduce the amount postgres
needs to fetch from its TOAST table.

use this for the get_all_notifications_for_service_for_csv view, as
_personalisation is not included in a csv row