Remove "Build ID" personalisation from FT fixtures
This was previously used to store the Jenkins build ID, but now the FTs
always just hardcode it to "No build ID".

This commit replaces it with a "name" field and a "content" field, which
is neater and allows both the functional tests and the API client
integration tests to use the same set of fixtures.

We use ((address_line_1)) in place of ((name)) for letter templates, to
match the existing fixtures for the API client integration tests.