made sqlalchemy database uri modifiable in test_downgrade
we were manually definining the SQLALCHEMY_DATABASE_URI in the script as
it needs to operate on a specific db name. However, this meant we were
ignoring any local config, which may contain noteworthy modifications to
the db URI: what URL/port it's on, or username/password details.

This commit changes so that if SQLALCHEMY_DATABASE_URI is not specific,
we'll use `postgresql://postgres:postgres@localhost/migration_test` as
before (but with explicit postgres:postgres username/password). If a
SQLALCHEMY_DATABASE_URI is passed in, then we'll take that postgres
server, but replace the database name with `migration_test` to ensure
we're not splatting over someone's regular or test notify database.