job tasks: introduce shatter_job_rows to create individual row tasks
in process_job and process_incomplete_job we're partly
bottlenecked by the rate at which we can put SQS messages
on the queue, causing the main job-processing task to
take longer than it should. this is a problem because it
holds a lock on the job row and it can sometimes get
killed by ECS scale-in halfway through processing, leaving
a zombie lock on the job row amongst possibly other things.

shatter_job_rows requires no database connection and its'
sole responsibility is to turn a sequence of task
args/kwargs pairs into individual tasks.

DEFAULT_SHATTER_JOB_ROWS_BATCH_SIZE is chosen with an eye
to avoid nearing the SQS message limit for these shatter
tasks.

this imports the "example equivalent" of itertools.batched
which is introduced in python 3.12. we should remove our
copy when we upgrade past that.