skip old letters in still_pending_virus_check task
previously, we were just selecting all letters more than 10 mins old
with status pending_virus_check. However, this involves checking
statuses all the way back through time - given this task runs every ten
minutes, we don't need to check the status of letters created a day ago
as they'll already have been fixed for example.

For now, just start with 30 minutes as the maximum time, in case the
scheduled task doesn't run for any reason.

If it doesn't run multiple times in a row, then we'd need to run the
task manually with a longer margin at least once - so I've made it an
optional parameter so we can manually trigger the task