Change provider balance on slow delivery every 5 minutes
Instead of every 10 minutes - this way we react faster, and
fewer text messages get delayed.

This becomes more important as we now send lots of login codes
for various organisations.

Here are some considerations for the change:
1. if we move from 10 minutes to 5 minutes, wouldn’t we use the notifications in our decision that were used in previous weighing too?

This is a valid question. This is happening already with current set up tho - as we look for last 15 minutes of sms, but we are changing weighing every 10 minutes.
So this wouldn’t be a new issue. Also this can be solved to either prioritise the best possible service for our users, or to prioritise fairest possible outcome for our providers. I believe getting text messages delivered promptly is more important than if a provider briefly loses more traffic because they were slow within last 15 minutes. Also if the slow down was just a short blip, on a subsequent delivery speed check some of the notifications will be from after the blip passed, and that should mean their priority has less chance of reducing further.

2. we should learn more before we merge it

On one side, fair, on the other, I kinda think it’s not a fundamental change, but rather a tuning, and we did give it some thought already, no?
We identified a problem (we adjust ratios too slowly when there is a period of slow delivery for one of our providers), we got an adjustment that should improve this, and we thought of potential flip-sides above (providers ratio will be lowered more quickly when they are slow, that means we will lower their ratio one more time than we would do before after their delivery speed recovers).