notifications _filter_query: omit status filter if "all" statuses are selected
avoid misleading the planner's estimates and avoid using
indexes it doesn't need to.

we don't have any NOTIFICATION_FAILED notifications in prod
anymore so don't require that to be specified to trigger this.

test is largely copied from get_notifications_for_job's tests,
with a bespoke addition to exercise this behaviour