dao_get_last_date_template_was_used: restrict notifications table search to midnight
beyond last midnight we should have a record of any notifications
in ft_notification_status, which should be faster to search (if
it isn't, we can address that easily). as long as we ensure we can
search notifications in created_at order (hence targeting it at a
specific index), we can avoid scanning old, cache-cold rows.

taking the template instance as an argument allows us to extract
all the field values we need to fill out these queries while
maintaining a convenient calling interface.