Change how we generate fake letter responses
We want to mark a letter as being delivered if it's sent using a test
key and the `SEND_LETTERS_ENABLED` flag is False. This was being done
by putting a fake response file in S3, but now that we are removing the
code related to response files, we can send a fake callback instead.
This lets us remove the rest of the code related to S3 response files.