Always send throught callback URL when sending letters
`LETTER_DELIVERY_CALLBACKS_ENABLED` is set to `True` in all
environments, so has been removed from the config. Since this is always
`True`, we can just always send the callback URL.

We still use `SEND_LETTERS_ENABLED` to determine whether or not we
should send the letters in the first place.