Add new task for generating fake letter responses
We want to mark a letter as being delivered if it's sent using a test
key and the `SEND_LETTERS_ENABLED` flag is False. This was being done
by putting a fake response file in S3, but now that we are removing the
code related to response files we can no longer use this method. This
adds a task to create a fake callback instead.

In later PRs, will then start using the new task and delete the code
related to response files.