Drop `daily_sorted_letter` table
We stopped writing to this table when we started receiving letter
delivery status via HTTPS callback instead of via response files in S3.
However, since 2023 we've been using the `ft_billing_letter_despatch`
table for letter billing and to store more detailed information about
which letters were classed as `sorted` vs `unsorted` so this data is not
needed.