always provide URI for psql
psql by default connects to a socket /var/run/postgresql/.s.PGSQL.5432
This checks the auth of the linux user running psql (in our docker
images, root) and tries to find a matching user in the database. We only
have a postgres database user however, so this fails in docker.

If we pass psql a URI that includes a username and password, it'll use
those to connect instead, allowing us to succeed