add handling of NOTIFY_EVENTLET_STATS
if we're using eventlet and NOTIFY_EVENTLET_STATS is set to 1,
we install the account_greenlet_times greenlet trace hook when
initializing the app

this should allow notifications_utils.eventlet's greenlet timing
and statistics functions to work and will annotate some of these
values onto app.request log messages.

this is a flag that should only be enabled when we're actively
tracking down an issue as there is probably a (minor) performance
penalty for calling this hook on every context switch.