delete inbound numbers and api keys
instead of calling `db.session.delete(service)`, call a lower-level
delete.

the db.session.delete operation looks at all the foreign key
relationships and if the service_id field is nullable on the child
object, then it just sets that to null - that's not what we want in this
case as we want to delete all those child artifacts. this forces us to
ensure that, for example, inbound_numbers are cleaned up properly

also delete the API keys a user has created, as the functional test user
creates an API key on the notify service that it uses to check 2fa codes