Row numbers in Returned Letters Report to match row numbers
in the spreadsheet with recipient data uploaded by service users
when they sent the letters.

This far, there has been a mismatch:

When our service users send letters in bulk ia Notify page, they
do so by uploading a spreadsheet with recipient data and persinalisation.
Each row in the spreadsheet becomes a letter.

In such a spreadsheet, first row is for column headers - for example
address_line_1, address_line_2, address_line_3, reference.
Then recipient rows start from row 2.

When we save the rows in notifications table in our db, we start
from index 0 for first recipient. That means that there is a
mismatch of 2 rows between our db and the spreadsheet that has been uploaded.

When collating data for Returned Letters Report, for some reason
we have been adding 1 to the db row number - which meant there was
still a mismatch - this time by one row.

This commit fixes the mismatch - from now on, the row number
we present in returned letters report will match the row number
in the uploaded spreadsheet.