Bump gunicorn to 23.0.0
The API is using a version of Gunicorn which is 3 years and 6 months
out of date.
https://github.com/benoitc/gunicorn/commit/1299ea9e967a61ae2edebe191082fd169b864c64

It has a high severity security vulnerability:
https://security.snyk.io/package/pip/gunicorn/21.0.0

Additionally, because we are referencing a dependency by SHA, Dependabot
can’t work out what version we are using and doesn’t flag this
vulnerability.

In order to deploy this we need to figure out:
- if we still expect the possibility of a performance regression
- how to test this