Update content of standard Notify templates
When spinning up a dev environment for user research, the content of the
Notify templates (such as the one telling you your service is now live)
is very out of date and doesn't match the current content.

This updates the templates so that they match production. This change
will only affect new environments, but saves us having to manually
update the content.