gunicorn: use psutil to emit post-request log output for long requests
this is ugly and nasty and should be removed as soon as we've
finished using it. we have to perform our own request timing
because we can't reasonably access the flask-level annotations
from the gunicorn side.

performing this from gunicorn has the advantage of being executed
*after* a response is sent, and so the 100ms delay this causes
(or unexpected errors this results in for that matter) shouldn't
prevent a user from getting their response.