gunicorn: use local copy of child_exit instead of importing from gds_metrics
this is something that contributes to the error splurge we get
every time our applications start - fixing this rules out the
possibility of the early-import-of-unpatched-io-routines being
related to our issues.

PROMETHEUS_MULTIPROC_DIR now set in entrypoint.sh to ensure this
is set early in docker and non-docker invocations. default-setting
from gds_metrics doesn't appear to happen early enough now.