Don't update free allowance for future years
We don't have annual_billing rows for future financial years, so don't
need to update them if a custom change is made to the free allowance.

Admin doesn't pass through a value for the financial year start when
updating the SMS free fragment limit, so it is only ever the current
year that we update.