Set the `has_custom_allowance` field when changing a free allowance
If the free allowance for a service has been changed (by a platform
admin user), as well as updating the free SMS fragments we should also
set the `has_custom_allowance` attribute to True.

The exception is if the free allowance is being changed to give it the
default for its organisation type - in this case the field should be
False since the service no longer has a custom allowance.