Persist custom free allowances from previous years
Previously, we were persisting a custom free allowance for a
service if it was 0 which had a couple of issues:
- services with a different custom allowance didn't have this persisted
  across years
- services which a free allowace of 0 due to being a high volume sender
  never had this reset if they stopped sending so many messages

We now automatically check if a service was a high volume sender in the
previous financial year and set the free allowance to 0. If a service
had a custom free allowance (unrelated to it being a high volume sender),
we now carry that forward to subsequent years.