Control sending weekly user research email via environment variable
We were always sending the weekly user research email, regardless of
environment. We had decided at the time that this was fine, but now that
we may have long-lived environments set up for user research we should
avoid sending the email to any of the participants if we forget to
archive them after the session.

This adds an environment variable to control whether or not the emails
should be sent, which is an approach we take for other tasks we only
want to happen in certain environments.