Delete test letter files from S3
When deleting the test letters from the Notifications table as part of
the `delete-notifications-for-service-and-type` task we weren't deleting
the test files from S3. We do delete the files in S3 for real letters, so
not deleting the files for test letters appears to have been an oversight.

This copies the approach we take for real letters and their files, and
deletes the test letter files in batches.