Shatter letter delivery tasks
When delivering letters to DVLA we are bottlenecked by the speed at
which we can create SQS messages. This is a problem because we have one
task which delivers all the letters, and it creates one SQS message at a
time. This task runs for a long time, and we can’t deploy while it’s
running.

This change makes it so that one new task (and SQS message) is created
for each 100 letters. Each of these new tasks – which can be picked up
asynchronously – then creates individual tasks for each of its 100
letters.

100 has been chosen as the biggest round number which fits comfortably
in the SQS message size limit.

The overall time taken will probably be about the same, but the maximum
time taken by any one task should be greatly reduced.

This is a similar approach taken by `shatter_jobs` in
https://github.com/alphagov/notifications-api/pull/4284/files