Bump Sentry SDK to 1.45.1
Fixes GHSA-g92j-qhmh-64v2

Not something that likely affects us, but clears down the only open
Dependabot warning on this repo.

This fix was released in sentry-sdk==2.8.0, then also backported to
sentry-sdk==1.45.1.

I think we could probably upgrade to Sentry >= 2 without making any
changes, but more testing would be needed to validate this. So just
going with the backport for now.

Specifying an exact version because that’s our convention.