Delete invalid letter PDFs as part of nightly job
We were relying on a lifecycle policy on the S3 bucket to delete the PDF
files for letters which had failed validation. However, this is a fixed
value (14 days) and services have variable data retention. We already
delete the files for valid letters according to the service data
retention. By adding the status of "validation-failed" to the list of
completed statuses (which it is, since it is a final state), these
letter files now get included in the ones that we delete overnight.