Persist sms notifications with job id even if number invalid
Save them with:
- permanent-failure status
- 0 billable units
- 0 as rate multiplier

Such notifications should not be sent.

Why do this:
Jobs are also validated in admin repository when they are uploaded.

If this has not happen, it most likely means that these jobs are either
scheduled jobs or jobs from emergency contact lists that were uploaded
before changes to our phone validation code.

And then notification persisting and sending happened later, after
we changed the code to a more strict version.

We don't want these jobs to be disrupted, even more so because
they could be related to business continuity. Thus, so they
process and complete correctly, and so that reports for them
can be seen and downloaded, we are more permissive here, instead
of throwing an error.