shatter_job_rows: recursively subdivide batches if SQS message too big
this introduces a wrapper function, _shatter_job_rows_with_subdivision
which will catch an exception caused by an oversized SQS message
and split the batch in half before trying again, recursing if it
too throws the same error.

uses this in both process_job and process_incomplete_jobs. in the
latter case, catch the exception raised when a batch can't be
subdivided any further, log an exception and continue with the
next incomplete job. this is unusual in that it's the only
exception we treat like this - perhaps we should be catching
more exceptions like this and proceeding with the next job(s)
to avoid a single bad job stopping all incomplete jobs being
re-processed, but in this specific case it became clear to me
that allowing this exception to stop everything would be a silly
behaviour.